check old reroute controller's state when new one is set
